+++ include/inviwo/core/datastructures/spatialdata.h
@@ -357,12 +357,12 @@ void StructuredGridEntity<N>::setDimensions(const Vector<N, unsigned int>& dimen
 template <unsigned int N>
 Matrix<N + 1, float> StructuredGridEntity<N>::getIndexMatrix() const {
     Matrix<N + 1, float> indexMatrix(1.0f);
-    for (int i = 0; i < N; ++i) {
+    for (size_t i = 0; i < N; ++i) {
         indexMatrix[i][i] = dimensions_[i];
     }
 
     // Offset to coordinates to center them in the middle of the texel/voxel.
-    for (int i = 0; i < N; i++) {
+    for (size_t i = 0; i < N; i++) {
         indexMatrix[N][i] = -0.5f;
     }
     return indexMatrix;
